Applied project to requirements
swift compiler version set to 5 (5.10)
deployment target set to iOS 15
deleted generated code/comments
default project files restructured
