+++ ShipsInfo/Application/AppDelegate.swift
@@ -1,0 +1,17 @@
+//
+//  AppDelegate.swift
+//  ShipsInfo
+//
+//  Created by Oleksandr Savchenko on 17.12.24.
+//
+
+import UIKit
+import CoreData
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        true
+    }
+}
+
