+++ ShipsInfo/Application/SceneDelegate.swift
@@ -1,0 +1,17 @@
+//
+//  SceneDelegate.swift
+//  ShipsInfo
+//
+//  Created by Oleksandr Savchenko on 17.12.24.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+    var window: UIWindow?
+    
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let _ = (scene as? UIWindowScene) else { return }
+    }
+}
+
