+++ ShipsInfo/Resources/Assets.xcassets/Contents.json
