+++ ShipsInfo/Resources/Info.plist
